requirements.txt
fastapi==0.111.0
llama-cpp-python==0.2.38
python-dotenv==1.0.1
slowapi==0.1.7
uvicorn==0.29.0
